Revert "refactor(api): isolate iframe browse service"
This reverts commit b5c438afc77f66426370addb4927008af98aacb9.
