+++ api/src/services/mission-browse/__tests__/index.test.ts
@@ -31,7 +31,7 @@ vi.mock("@/services/mission", () => ({
   missionService: { findMissionsByIds: findMissionsByIdsMock, findOneMissionBy: findOneMissionByMock },
 }));
 
-import { missionBrowseService } from "@/services/mission-browse";
+import { MissionBrowseIndexUnavailableError, missionBrowseService } from "@/services/mission-browse";
 
 const baseParams = { page: 1, pageSize: 20, diffuseurPublisherId: "diffuseur-1" };
 // Filtre de diffusion attendu : le snapshot `mission_diffusion` dénormalisé dans le document Typesense.
@@ -115,6 +115,84 @@ describe("missionBrowseService.browse", () => {
 
     expect(facets.type_mission).toEqual([{ key: "benevolat", count: 42 }]);
     expect(facets.domaine).toEqual([]);
+  });
+
+  it("utilise la recherche, le tri et les facettes du widget sans modifier le mode intégrateur", async () => {
+    multiSearchMock.mockResolvedValue([
+      { hits: [], found: 0 },
+      { hits: [], found: 0, facet_counts: [{ field_name: "mission_domain", counts: [] }] },
+      { hits: [], found: 0, facet_counts: [{ field_name: "publisherOrganizationFacet", counts: [{ value: "org-1|||Organisation 1", count: 3 }] }] },
+      { hits: [], found: 0, facet_counts: [{ field_name: "departmentNames", counts: [] }] },
+      { hits: [], found: 0, facet_counts: [{ field_name: "remote", counts: [] }] },
+      { hits: [], found: 0, facet_counts: [{ field_name: "countryCodes", counts: [] }] },
+      { hits: [], found: 0, facet_counts: [{ field_name: "schedule", counts: [] }] },
+      { hits: [], found: 0, facet_counts: [{ field_name: "tasks", counts: [] }] },
+      { hits: [], found: 0, facet_counts: [{ field_name: "audience", counts: [] }] },
+      { hits: [], found: 0, facet_counts: [{ field_name: "openToMinors", counts: [{ value: "true", count: 2 }] }] },
+      {
+        hits: [],
+        found: 0,
+        facet_counts: [
+          { field_name: "reducedMobilityAccessible", counts: [{ value: "true", count: 1 }] },
+          { field_name: "closeToTransport", counts: [{ value: "true", count: 4 }] },
+        ],
+      },
+    ]);
+
+    const result = await missionBrowseService.browse({
+      ...baseParams,
+      widgetMode: true,
+      baseFilterBy: "publisherId:=`publisher-1`",
+      search: "solidarité",
+    });
+
+    expect(resultsSearch()).toMatchObject({
+      q: "solidarité",
+      query_by: "title,publisherOrganizationFacet,cityNames,mission_domain",
+      sort_by: "startAt:desc,createdAt:desc",
+      filter_by: "publisherId:=`publisher-1`",
+    });
+    expect(result.facets.organization).toEqual([{ key: "org-1", label: "Organisation 1", count: 3 }]);
+    expect(result.facets.minor).toEqual([{ key: "yes", count: 2 }]);
+    expect(result.facets.accessibility).toEqual([
+      { key: "reducedMobilityAccessible", count: 1 },
+      { key: "closeToTransport", count: 4 },
+    ]);
+  });
+
+  it("conserve les missions locales hors rayon dans une recherche géolocalisée", async () => {
+    await missionBrowseService.browse({
+      ...baseParams,
+      widgetMode: true,
+      baseFilterBy: "publisherId:=`publisher-1`",
+      lat: 48.8566,
+      lon: 2.3522,
+      distanceKm: 50,
+      remote: ["no", "local"],
+    });
+
+    expect(resultsSearch().filter_by).toContain("(locations:(48.8566,2.3522,50 km) || remote:=`local`)");
+  });
+
+  it("impose le rayon quand le filtre remote exclut les missions locales", async () => {
+    await missionBrowseService.browse({
+      ...baseParams,
+      widgetMode: true,
+      baseFilterBy: "publisherId:=`publisher-1`",
+      lat: 48.8566,
+      lon: 2.3522,
+      distanceKm: 50,
+      remote: ["full", "possible"],
+    });
+
+    expect(resultsSearch().filter_by).toContain("locations:(48.8566,2.3522,50 km)");
+    expect(resultsSearch().filter_by).not.toContain("|| remote:=`local`");
+  });
+
+  it("expose une erreur métier lorsque Typesense est indisponible", async () => {
+    multiSearchMock.mockRejectedValue(new Error("Typesense unavailable"));
+
+    await expect(missionBrowseService.browse(baseParams)).rejects.toBeInstanceOf(MissionBrowseIndexUnavailableError);
   });
 
   it("restreint le détail au snapshot matérialisé du diffuseur", async () => {
