+++ api/tests/integration/api/endpoints/iframe/browse.test.ts
@@ -3,16 +3,16 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 
 const browseMock = vi.hoisted(() => vi.fn());
 const buildWidgetBaseFilterMock = vi.hoisted(() => vi.fn());
-const IframeBrowseIndexUnavailableErrorMock = vi.hoisted(() => class IframeBrowseIndexUnavailableError extends Error {});
+const MissionBrowseIndexUnavailableErrorMock = vi.hoisted(() => class MissionBrowseIndexUnavailableError extends Error {});
 
-vi.mock("@/services/iframe-browse", () => ({
-  IframeBrowseIndexUnavailableError: IframeBrowseIndexUnavailableErrorMock,
-  iframeBrowseService: {
+vi.mock("@/services/mission-browse", () => ({
+  MissionBrowseIndexUnavailableError: MissionBrowseIndexUnavailableErrorMock,
+  missionBrowseService: {
     browse: browseMock,
   },
 }));
 
-vi.mock("@/services/iframe-browse/widget-filters", () => ({
+vi.mock("@/services/mission-browse/widget-filters", () => ({
   buildWidgetBaseFilter: buildWidgetBaseFilterMock,
 }));
 
@@ -55,7 +55,9 @@ describe("GET /iframe/:id/browse", () => {
       expect.objectContaining({
         search: "solidarité",
         remote: ["full", "possible"],
+        diffuseurPublisherId: publisher.id,
         baseFilterBy: "widget-filter",
+        widgetMode: true,
         moderatedBy: PUBLISHER_IDS.JEVEUXAIDER,
         page: 3,
         pageSize: 5,
@@ -91,7 +93,7 @@ describe("GET /iframe/:id/browse", () => {
 
   it("retourne 503 lorsque Typesense est indisponible", async () => {
     const widget = await createTestWidget();
-    browseMock.mockRejectedValue(new IframeBrowseIndexUnavailableErrorMock());
+    browseMock.mockRejectedValue(new MissionBrowseIndexUnavailableErrorMock());
 
     await request(app).get(`/iframe/${widget.id}/browse`).expect(503);
   });
